+++ app/Main.hs
@@ -3,15 +3,15 @@ module Main (main) where
 import           ClassyPrelude
 
 import qualified Data.ByteString.Char8 as BC
-import           Options.Applicative   (execParser, flag', fullDesc, header, help, helper,
-                                        info, long, progDesc, short, showDefault,
-                                        strOption, value, (<**>))
+import           Options.Applicative   (auto, execParser, flag', fullDesc, header, help,
+                                        helper, info, long, option, progDesc, short,
+                                        showDefault, strOption, value, (<**>))
 import           System.Directory      (doesFileExist)
 
 
 import           Template              (spiralWorldTemplate, zigZagWorldTemplate)
 import           World                 (addCoords, findLargestIsland, mkWorld,
-                                        mkWorldByTemplate, parseWorldConcurr)
+                                        mkWorldByTemplate, parseWorld)
 
 -- | Scroll can be processed in these ways, which should be specified in cmd line
 data ProcessType = Spiral     -- ^ use SpiralWorld template
@@ -24,6 +24,7 @@ data ProcessType = Spiral     -- ^ use SpiralWorld template
 data Args = Args
   { argScrollPath  :: FilePath    -- ^ config file to use
   , argProcessType :: ProcessType -- ^ how to process the scroll
+  , argChunkSize   :: Int         -- ^ size of the chunk to perform concurrent parse
   } deriving (Show)
 
 -- | Entry point
@@ -37,7 +38,7 @@ main = do
   doesFileExist argScrollPath >>= \case
     False -> putStrLn noFileError
     True  -> do
-      pw <- parseWorldConcurr =<< BC.readFile argScrollPath
+      pw <- parseWorld argChunkSize =<< BC.readFile argScrollPath
 
       let vlgs = case argProcessType of
                    Spiral     -> addCoords $ mkWorldByTemplate spiralWorldTemplate pw
@@ -54,6 +55,9 @@ main = do
                             <> value "data/the.scroll" <> showDefault
                             <> help "Path of the scroll file")
                 <*> (spiralP <|> zigZagP <|> noTemplateP)
+                <*> option auto (long "chunk_size" <> short 'z'
+                              <> value 3000 <> showDefault
+                              <> help "Size of chunk in bytes for concurrent parse")
 
     spiralP     = flag' Spiral     (long "spiral"      <> help "Spiral template")
     zigZagP     = flag' ZigZag     (long "zig-zag"     <> help "Zig-zag template")