+++ src/Template.hs
@@ -6,7 +6,6 @@ module Template
   , pointCoordInSpiral
   , spiralWorldTemplate
   , zigZagWorldTemplate
-
   )
 where
 
@@ -44,31 +43,32 @@ mapOver s = map (BC.pack . map go) . unTemplate
 -- | Get the coordinates for a sequential point in an n-dimensional 2D square spiral.
 --   Coord system is top-left based starting with (1,1)
 pointCoordInSpiral :: Int -> Int -> Coord
-pointCoordInSpiral dimension srch =
-  if srch > pow2 dimension
+pointCoordInSpiral 1 1 = (1,1)
+pointCoordInSpiral 2 1 = (1,1)
+pointCoordInSpiral 2 2 = (2,1)
+pointCoordInSpiral 2 3 = (2,2)
+pointCoordInSpiral 2 4 = (1,2)
+pointCoordInSpiral dimension position = do
+  let fullDim = floor   @Float . sqrt $ fromIntegral position
+      jumpDim = ceiling @Float $ fromIntegral (dimension - fullDim) / 2
+      initPos = if even fullDim
+                  then (1 + jumpDim, fullDim + jumpDim)
+                  else (fullDim + jumpDim, 1 + jumpDim)
+      newPos  = position - pow2 fullDim
+
+  if position >= pow2 dimension
     then error "Matrix dimension too small for searched point"
-    else go 1 srch (middlePoint, middlePoint)
+    else go (fullDim + 1) newPos initPos
   where
-    middlePoint     = ceiling @Float $ fromIntegral dimension / 2
-    go !dim !n (!x, !y) = do
-      if | n <= 0           -> (x, y)
-         | dim == 1         -> nextDim subN (x,y)
-         | n >= dimDiffArea -> nextDim (subN - (dim - 1) * 2) fullDim
-         | even dim         -> nextDim partialN (x + 1 - xMove, y + yMove)
-         | odd  dim         -> nextDim partialN (x - 1 + xMove, y - yMove)
-         | otherwise        -> (x, y) -- should never be reached
+    go !dim n (!x, !y) =
+      if | n <= 0    -> (x, y)
+         | even dim  -> (x + 1 - xMove, y + yMove)
+         | odd  dim  -> (x - 1 + xMove, y - yMove)
+         | otherwise -> error "should never be reached"
 
       where
-        maxMoves         = dim - 1
-        subN             = n - 1
-        yMove            = min maxMoves subN
-        xMove            = min maxMoves (subN - yMove)
-        partialN         = subN - xMove - yMove
-        dimDiffArea      = pow2 dim - pow2 (dim-1)
-        nextDim !n' !pos = go (dim + 1) n' pos
-        fullDim          = if even dim
-                             then (x + 1 - maxMoves, y + maxMoves)
-                             else (x - 1 + maxMoves, y - maxMoves)
+        yMove = min (dim - 1) (n - 1)
+        xMove = min (dim - 1) (n - 1 - yMove)
 
 -- | Create a square spiral template starting in the middle and going clock-wise
 spiralWorldTemplate :: Int -> Template
@@ -138,8 +138,29 @@ pow2 = (^ (2 :: Int))
 --      , [36,35,34,33,32,31]
 --      ]
 
+-- m7 :: [[Int]]
+-- m7 = [ [43,44,45,46,47,48,49]   -- R DDDDDD LLLLLL
+--      , [42,21,22,23,24,25,26]
+--      , [41,20, 7, 8, 9,10,27]
+--      , [40,19, 6, 1, 2,11,28]
+--      , [39,18, 5, 4, 3,12,29]
+--      , [38,17,16,15,14,13,30]
+--      , [37,36,35,34,33,32,31]
+--      ]
+
+-- m8 :: [[Int]]
+-- m8 = [ [43,44,45,46,47,48,49,50]   -- R DDDDDDD LLLLLLL
+--      , [42,21,22,23,24,25,26,51]
+--      , [41,20, 7, 8, 9,10,27,52]
+--      , [40,19, 6, 1, 2,11,28,53]
+--      , [39,18, 5, 4, 3,12,29,54]
+--      , [38,17,16,15,14,13,30,55]
+--      , [37,36,35,34,33,32,31,56]
+--      , [64,63,62,61,60,59,58,57]
+--      ]
+
 -- m9 :: [[Int]]
--- m9 = [ [73,74,75,76,77,78,79,81,82]  -- R DDDDDDDD LLLLLLLL
+-- m9 = [ [73,74,75,76,77,78,79,80,81]  -- R DDDDDDDD LLLLLLLL
 --      , [72,43,44,45,46,47,48,49,50]
 --      , [71,42,21,22,23,24,25,26,51]
 --      , [70,41,20, 7, 8, 9,10,27,52]