+++ ChangeLog.md
@@ -1,5 +1,15 @@
 # Changelog for artificial-haskell-test
 
+
+## 0.0.0.3
+  Speed improvements (234 ms -> 112 ms):
+  - Map -> HashMap
+  - skip converting ByteString <-> String
+  - do not use BC.cons - expensive
+  - remove sequential parse
+  - configurable chunk size for concurrent parse
+  - optimise direct coordination find to have similar speed regardless of search position
+
 ## 0.0.0.2
   - Fix mempry leak and CPU choke
   - Add direct village coordinate search